requirements.txt
langchain
openai
streamlit
jupyter
tiktoken
# faiss-cpu
python-dotenv
ipython
streamlit_chat
sentence-transformers
# chromadb
unstructured
pinecone-client
pypdf
huggingface-hub